Fixed a memory leak, saving the camera position in the config, fixed saving lighting, added 360° batch rendering, model rotation control in the viewport, resizing the viewport
Fixed a memory leak that was caused by console logs.
Added saving the camera position between sessions to the config file.
Fixed saving the lighting, as it was not applied after a new launch, but simply displayed the coordinates and intensity.
Added a change in the rotation of the object on the stage.
Then I added automatic batch rendering with a step that is set in the interface, and added resizing the viewport, since 1:1 sprites are not needed everywhere.